Split warehouse tests
