+++ saleor/graphql/warehouse/tests/queries/test_warehouse.py
@@ -1,0 +1,203 @@
+import graphene
+
+from ....tests.utils import (
+    assert_no_permission,
+    get_graphql_content,
+    get_graphql_content_from_response,
+)
+
+QUERY_WAREHOUSE = """
+query warehouse($id: ID!){
+    warehouse(id: $id) {
+        id
+        name
+        companyName
+        email
+        shippingZones(first: 100) {
+            edges {
+                node {
+                    name
+                    countries {
+                        country
+                    }
+                }
+            }
+        }
+        address {
+            streetAddress1
+            streetAddress2
+            postalCode
+            city
+            phone
+        }
+    }
+}
+"""
+
+
+def test_warehouse_query(staff_api_client, warehouse, permission_manage_products):
+    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
+
+    response = staff_api_client.post_graphql(
+        QUERY_WAREHOUSE,
+        variables={"id": warehouse_id},
+        permissions=[permission_manage_products],
+    )
+    content = get_graphql_content(response)
+
+    queried_warehouse = content["data"]["warehouse"]
+    assert queried_warehouse["name"] == warehouse.name
+    assert queried_warehouse["email"] == warehouse.email
+
+    shipping_zones = queried_warehouse["shippingZones"]["edges"]
+    assert len(shipping_zones) == warehouse.shipping_zones.count()
+    queried_shipping_zone = shipping_zones[0]["node"]
+    shipipng_zone = warehouse.shipping_zones.first()
+    assert queried_shipping_zone["name"] == shipipng_zone.name
+    assert len(queried_shipping_zone["countries"]) == len(shipipng_zone.countries)
+
+    address = warehouse.address
+    queried_address = queried_warehouse["address"]
+    assert queried_address["streetAddress1"] == address.street_address_1
+    assert queried_address["postalCode"] == address.postal_code
+
+
+def test_warehouse_query_as_staff_with_manage_orders(
+    staff_api_client, warehouse, permission_manage_orders
+):
+    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
+
+    response = staff_api_client.post_graphql(
+        QUERY_WAREHOUSE,
+        variables={"id": warehouse_id},
+        permissions=[permission_manage_orders],
+    )
+    content = get_graphql_content(response)
+
+    queried_warehouse = content["data"]["warehouse"]
+    assert queried_warehouse["name"] == warehouse.name
+    assert queried_warehouse["email"] == warehouse.email
+
+    shipping_zones = queried_warehouse["shippingZones"]["edges"]
+    assert len(shipping_zones) == warehouse.shipping_zones.count()
+    queried_shipping_zone = shipping_zones[0]["node"]
+    shipipng_zone = warehouse.shipping_zones.first()
+    assert queried_shipping_zone["name"] == shipipng_zone.name
+    assert len(queried_shipping_zone["countries"]) == len(shipipng_zone.countries)
+
+    address = warehouse.address
+    queried_address = queried_warehouse["address"]
+    assert queried_address["streetAddress1"] == address.street_address_1
+    assert queried_address["postalCode"] == address.postal_code
+
+
+def test_warehouse_query_as_staff_with_manage_shipping(
+    staff_api_client, warehouse, permission_manage_shipping
+):
+    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
+
+    response = staff_api_client.post_graphql(
+        QUERY_WAREHOUSE,
+        variables={"id": warehouse_id},
+        permissions=[permission_manage_shipping],
+    )
+    content = get_graphql_content(response)
+
+    queried_warehouse = content["data"]["warehouse"]
+    assert queried_warehouse["name"] == warehouse.name
+    assert queried_warehouse["email"] == warehouse.email
+
+    shipping_zones = queried_warehouse["shippingZones"]["edges"]
+    assert len(shipping_zones) == warehouse.shipping_zones.count()
+    queried_shipping_zone = shipping_zones[0]["node"]
+    shipipng_zone = warehouse.shipping_zones.first()
+    assert queried_shipping_zone["name"] == shipipng_zone.name
+    assert len(queried_shipping_zone["countries"]) == len(shipipng_zone.countries)
+
+    address = warehouse.address
+    queried_address = queried_warehouse["address"]
+    assert queried_address["streetAddress1"] == address.street_address_1
+    assert queried_address["postalCode"] == address.postal_code
+
+
+def test_warehouse_query_as_staff_with_manage_apps(
+    staff_api_client, warehouse, permission_manage_apps
+):
+    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
+
+    response = staff_api_client.post_graphql(
+        QUERY_WAREHOUSE,
+        variables={"id": warehouse_id},
+        permissions=[permission_manage_apps],
+    )
+
+    assert_no_permission(response)
+
+
+def test_warehouse_query_as_customer(user_api_client, warehouse):
+    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
+
+    response = user_api_client.post_graphql(
+        QUERY_WAREHOUSE,
+        variables={"id": warehouse_id},
+    )
+
+    assert_no_permission(response)
+
+
+def test_staff_query_warehouse_by_invalid_id(
+    staff_api_client, warehouse, permission_manage_shipping
+):
+    id = "bh/"
+    variables = {"id": id}
+    response = staff_api_client.post_graphql(
+        QUERY_WAREHOUSE, variables, permissions=[permission_manage_shipping]
+    )
+    content = get_graphql_content_from_response(response)
+    assert len(content["errors"]) == 1
+    assert content["errors"][0]["message"] == f"Couldn't resolve id: {id}."
+    assert content["data"]["warehouse"] is None
+
+
+def test_staff_query_warehouse_with_invalid_object_type(
+    staff_api_client, permission_manage_shipping, warehouse
+):
+    variables = {"id": graphene.Node.to_global_id("Order", warehouse.pk)}
+    response = staff_api_client.post_graphql(
+        QUERY_WAREHOUSE, variables, permissions=[permission_manage_shipping]
+    )
+    content = get_graphql_content(response)
+    assert content["data"]["warehouse"] is None
+
+
+QUERY_WAREHOUSE_BY_EXTERNAL_REFERENCE = """
+query warehouse($id: ID, $externalReference: String){
+    warehouse(id: $id, externalReference: $externalReference) {
+        id
+        externalReference
+    }
+}
+"""
+
+
+def test_warehouse_query_by_external_reference(
+    staff_api_client, warehouse, permission_manage_products
+):
+    # given
+    ext_ref = "test-ext-ref"
+    warehouse.external_reference = ext_ref
+    warehouse.save(update_fields=["external_reference"])
+    variables = {"externalReference": ext_ref}
+
+    # when
+    response = staff_api_client.post_graphql(
+        QUERY_WAREHOUSE_BY_EXTERNAL_REFERENCE,
+        variables=variables,
+        permissions=[permission_manage_products],
+    )
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["warehouse"]
+    assert data["externalReference"] == ext_ref
+    assert data["id"] == graphene.Node.to_global_id("Warehouse", warehouse.id)
